cleanup - revert some commits from earlier
