+++ apps/gitness/src/routes.tsx
@@ -10,8 +10,8 @@ import {
 } from '@harnessio/ui/views'
 
 import { AppShell } from './components-v2/app-shell'
-import { AppShellMFE } from './components-v2/app-shell-mfe'
 import { ProjectDropdown } from './components-v2/breadcrumbs/project-dropdown'
+import { AppShellMFE } from './components-v2/mfe/app-shell.tsx'
 import { AppProvider } from './framework/context/AppContext'
 import { AppRouterProvider } from './framework/context/AppRouterProvider'
 import { ExplorerPathsProvider } from './framework/context/ExplorerPathsContext'
